elite pass: xtask, benchmarks, channel hopping, deny + release workflows
The infrastructure a senior reviewer would want to see before merging
a project into their fleet.

Code quality:
- #![forbid(unsafe_code)] on every lib and binary root
- cargo clippy -D warnings clean under both default and --features live
- LiveCapture::next renamed to next_packet to match pcap::Capture and
  dodge the std::iter::Iterator::next naming clash
- FrameSource trait + every call site renamed to match
- 8 pre-existing clippy nits swept (redundant guards, useless format,
  needless_option_as_deref, unnecessary casts, doc list overindents,
  should_implement_trait, needless_return)

Real feature work:
- Channel hopping for airodump via a ChannelControl trait, iw-based
  Linux impl, deterministic Hopper walker, and a `--hop` flag that
  accepts 24ghz / 5ghz / 6ghz / comma-separated lists
- Current channel rendered live next to the TUI counters

Testing:
- wifi/tests/fuzz_like.rs: 15k pseudorandom buffers through both
  parsers with a hard no-panic requirement, plus end-to-end builder
  round-trips and a pcap writer/reader round-trip. 12 new integration
  tests on top of the 8 unit tests.
- wifi/benches/parser.rs: criterion benchmarks for the 802.11 parser
  (single beacon, 1k-frame mixed workload, radiotap strip + parse).
  Throughput metrics land under target/criterion.

Tooling:
- xtask/ crate with sample / ci / dist / completions / manpages
  commands; hooked via a cargo alias so `cargo xtask ci` just works
- Shell completions for bash/zsh/fish/PowerShell/elvish via
  clap_complete
- Man pages via clap_mangen
- justfile aliasing the same entry points for contributors that
  prefer `just`
- deny.toml gating licenses and advisories
- .github/dependabot.yml (weekly, grouped patch+minor)
- .github/workflows/audit.yml (cargo-audit + cargo-deny on push,
  PR, and a weekly cron)
- .github/workflows/release.yml: cross-platform tagged-release build
  for x86_64 linux/macos/windows-msvc and aarch64 macos, uploads
  tar.gz/zip + SHA-256 to GitHub Releases

Community / repo hygiene:
- CI now runs `cargo clippy -D warnings` + an MSRV 1.75 check + a
  bench-compile job in addition to the fmt/check/test matrix
- CONTRIBUTING.md, CODE_OF_CONDUCT.md, SECURITY.md
- .github/ISSUE_TEMPLATE/{bug_report,feature_request,config}.yml
- .github/pull_request_template.md
- Expanded README with the new "development" section documenting
  xtask, just, CI jobs, and the unsafe-forbid + audit badges
- CHANGELOG.md [Unreleased] entry documenting all of the above

+++ airbase/src/main.rs
@@ -10,6 +10,8 @@
 //! Many APs means many beacons per interval, so the tool walks through
 //! the list inside a ticker.
 
+#![forbid(unsafe_code)]
+
 use std::time::Duration;
 
 use airscope_core::MacAddr;
@@ -145,8 +147,8 @@ mod hex {
     pub fn encode(data: &[u8]) -> String {
         let mut out = String::with_capacity(data.len() * 2);
         for b in data {
-            out.push(to_hex((b >> 4) as u8));
-            out.push(to_hex((b & 0x0f) as u8));
+            out.push(to_hex(b >> 4));
+            out.push(to_hex(b & 0x0f));
         }
         out
     }
